added php files
added a php version of index.html & login.html
added functions for the php files 
login.php still needs work: login cant seem to be logging in.
register.php is functional, users can create their account and their info is available on PHPMyAdmin database. After creating an account, the user is redirected to the login page. 
connection.php is functional and checks whether the web app is able to connect to PHPMyAdmin database.